# multiple small fixes