--- a/.editorconfig
+++ b/.editorconfig
@@ -6,7 +6,7 @@ root = true
 
 [*]
 # Change these settings to your own preference
-indent_style = space
+indent_style = tab
 indent_size = 2
 
 # We recommend you to keep these unchanged