- `gidFromG` catches segfaults now, as comment suggests it should
- added `FastGetGoIDAvailable` to tell if fast version of `GetGoID` is available
- changed spaces to tabs in the `.editorconfig`
- `go fmt`

There is no significant performance difference.

Before:
```
 ❯ go test -bench .
goos: darwin
goarch: arm64
pkg: github.com/observeinc/goid
BenchmarkSlowGid-10    	  636439	      1883 ns/op	      64 B/op	       2 allocs/op
BenchmarkFastGid-10    	592396588	         2.021 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetGoID-10    	587365522	         2.036 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/observeinc/goid	5.134s
```

After:
```
 ❯ go test -bench .
goos: darwin
goarch: arm64
pkg: github.com/observeinc/goid
BenchmarkSlowGid-10    	  638582	      1864 ns/op	      64 B/op	       2 allocs/op
BenchmarkFastGid-10    	595019316	         2.014 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetGoID-10    	588988731	         2.036 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/observeinc/goid	5.046s
```